+++ RELEASE_NOTES.md
@@ -7,6 +7,7 @@ This is the offical release of vpinfe v2.0.3
 - More managerUI clearing/seperation.  Logs now on its own page.  added vpx logs
 - QRCode in the mainmenu for linking to remote.
 - VPinPlay Player QR code.
+- VPinPlay Multiple player profiles
 
 ### Fixes
 None